ppc/xive: introduce the XIVE Event Queues
The Event Queue Descriptor (EQD) table is an internal table of the
XIVE routing sub-engine, the IVRE. It specifies on which Event Queue
the Event Queue data should be posted when an exception occurs (and
later on pulled by the OS) and which Virtual Processor to notify.

The routing algorithm of the XiveRouter is extended to push in the
system memory event queue the Event Queue data defined in the
associated IVE. The event queue is a circular buffer in RAM provided
by the OS, one per server and priority couple, containing Event Queue
entries. These are 4 bytes long, the first bit being a 'generation'
bit and the 31 following bits the EQ Data field.

The EQ Data field is a way to set an invariant logical event source
number for an IRQ. It is set with the H_INT_SET_SOURCE_CONFIG hcall
when the EISN flag is used.

Signed-off-by: Cédric Le Goater <clg@kaod.org>