ppc/xive: introduce a simplified XIVE presenter
The last sub-engine of the XIVE architecture is the Interrupt
Virtualization Presentation Engine (IVPE). On HW, they share elements,
the Power Bus interface (CQ), the routing table descriptors, and they
can be combined in the same HW logic. We do the same in QEMU and
combine both engines in the XiveRouter for simplicity.

When the IVRE has completed its job of matching an event source with a
Virtual Processor (VP) target to notify, it forwards the event
notification to the IVPE handling the Virtual Processor threads. The
IVPE scans the thread interrupt context of the VPs dispatched on the
HW processor threads and if a match is found, it signals the
thread. If not, the IVRE escalates the notification to some other
targets and records the notification in a backlog queue.

The IVPE maintains the thread interrupt context state for each of its
VPs not dispatched on HW processor threads in the Virtual Processor
Descriptor (VPD) table.

The model currently only supports single VP notifications.

Signed-off-by: Cédric Le Goater <clg@kaod.org>