ppc/xive: add support for the EQ Event State buffers
The Event Queue Descriptor also contains two Event State Buffers
providing further coalescing of interrupts, one for the notification
event (ESn) and one for the escalation events (ESe). A MMIO page is
assigned for each to control the EOI through loads only. Stores are
not allowed.

The EQ ESB are modeled through an object resembling the 'XiveSource'
It is stateless as the EQ state bits are backed into the XiveEQ
structure under the XiveRouter and the MMIO accesses follow the same
rules as the standard source ESBs.

EQ ESBs are not supported by the Linux drivers neither on OPAL nor on
sPAPR. Nnevetherless, it provides a mean to study the question in the
future and validates a bit more the XIVE model.

Signed-off-by: Cédric Le Goater <clg@kaod.org>