spapr/xive: use the VCPU id as a VP identifier in the OS CAM.
For the IVPE to find a matching VP among the VPs dispatched on the
physical processor threads, the model needs to update the OS CAM line
of the XIVE thread interrupt context with the VP identifier.

The model uses the VCPU id as the VP identifier for the sPAPR and
provide a set of helpers to do the conversion between identifiers.
EQ ids are also derived from the VCPU id.

The sPAPRXive model does not need a VPD table. Nevertheless, the
XiveRouter VP handlers provide some extra checks for the routing
algorithm.

Signed-off-by: Cédric Le Goater <clg@kaod.org>