spapr/xive: use the VCPU id as a VP identifier
For the IVPE to find a matching VP among the VPs dispatched on the
physical processor threads, the OS CAM line of the XIVE thread
interrupt context should contain the identifier of the targeted VP. On
a real system, this is updated by the hypervisor when a VP is
scheduled to run on a HW thread. Under QEMU, the model emulates the
same behavior by hardwiring the VP identifier in the thread context
registers at reset.

The VP identifier used by the sPAPRXive model is the VCPU id. The EQ
identifier is also derived from the VCPU id. A set of helpers doing
the conversion between identifiers are provided for the hcalls
configuring the sources and the EQs.

The model does not need a VPD table but the provided VP operations
perform some extra checks for the routing algorithm.

Signed-off-by: Cédric Le Goater <clg@kaod.org>