spapr/xive: add migration support for KVM
This extends the KVM XIVE models to handle the state synchronization
with KVM, for the monitor usage and for the migration.

The migration priority of the sPAPRXive KVM model is raised to operate
first and orchestrate the capture sequence of all XIVE models. All
sources are masked to quiesce the interrupt flow and a XIVE xync is
performed to stabilize the OS EQs. The state of the EQDs are then
captured, followed by the state of the thread contexts. When done, a
rollback is performed to restore the sources to their initial state.

Post_load is simpler and only needs to restore the different states in
the correct order. The SPAPRXIVE model first restores the routing
tables. Then the sources and thread contexts come in play

The get/set operations rely on their KVM counterpart in the host
kernel which acts as a proxy for OPAL, the host firmware. Extra
quiescence points are possibly needed. Stress tests will say.

Signed-off-by: Cédric Le Goater <clg@kaod.org>